Move FRC implementation to device server module
